# manoward/static/sw/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Manowar 2.x API",
    "description": "Manowar 2 API. Designed to be a progrmattic way to interact with the data\ncollected and analyzed by Jellyfish 2.\n",
    "version": "2.6"
  },
  "servers": [
    {
      "url": "/manowar/v2"
    }
  ],
  "components": {
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic"
      }
    }
  },
  "security": [
    {
      "basicAuth": []
    }
  ],
  "paths": {
...
    "/auditresults/{audit_id}/{request_timestamp}/": {
      "get": {
        "description": "Designed to grab a list of hosts that either pass or fail an audit\nalong with the relevant data about each host. Similar to the auditresults\nendpoint, except that this one also takes a timestamp so that it can\nreturn the results as it appears at a particular time.\n",
        "tags": [
          "audits"
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "parameters": [
          {
            "name": "audit_id",
            "in": "path",
            "description": "The id of the audit you wish to get hosts back for. Needs to be speicied here or\nin the query string. This parameter is not technically required.\n",
            "schema": {
              "type": "integer"
            },
            "required": true
          },
          {
            "name": "request_timestamp",
            "in": "path",
            "description": "A Unix Timestamp that you want to check for the date against.\nMust be a positive integer. This variable is required. If you\nwish to get the \"latest\" you shoul utilize the /auditresults\nendpoint.\n",
            "schema": {
              "type": "integer"
            },
            "required": true
          },
          {
            "name": "bucket",
            "in": "query",
            "description": "A regex to match for the bucket. [PCRE](https://mariadb.com/kb/en/mariadb/regexp/) type\nregular expressions are accepted. Matched on the status column in the audits_by_host table. With exact\nwill match exactly.\n",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "auditResult",
            "in": "query",
            "description": "A regex to match for the Audit Result (pass/fail). [PCRE](https://mariadb.com/kb/en/mariadb/regexp/) type\nregular expressions are accepted. Matched on the audit_result column in the audit_by_host table. With exact\nwill match exactly.\n",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "pass",
                "fail",
                "notafflicted"
              ]
            }
          },
          {
            "name": "auditResultText",
            "in": "query",
            "description": "A regex to match for the Audit Result Text. [PCRE](https://mariadb.com/kb/en/mariadb/regexp/) type\nregular expressions are accepted. Matched on the audit_result_text column in the audit_by_host table. With exact\nwill match exactly.\n",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "hostname",
            "in": "query",
            "description": "A regex to match for the hostname. [PCRE](https://mariadb.com/kb/en/mariadb/regexp/) type\nregular expressions are accepted. Matched on the hostname column in the host table. With exact\nwill match exactly.\n",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "A regex to match for the status. [PCRE](https://mariadb.com/kb/en/mariadb/regexp/) type\nregular expressions are accepted. Matched on the status column in the host table. With exact\nwill match exactly.\n",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pop",
            "in": "query",
            "description": "A regex to match for the pop. [PCRE](https://mariadb.com/kb/en/mariadb/regexp/) type\nregular expressions are accepted. Matched on the status column in the host table. With exact\nwill match exactly.\n",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "srvtype",
            "in": "query",
            "description": "A regex to match for the srvtype. [PCRE](https://mariadb.com/kb/en/mariadb/regexp/) type\nregular expressions are accepted. Matched on the status column in the host table. With exact\nwill match exactly.\n",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "resource",
            "in": "query",
            "description": "Matches the MOWN Resource either via Regex or Exactly (with exact). Keep in mind resources include\nthe resource qualifier and resource subtype if applicable.\n",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "partition",
            "in": "query",
            "description": "Matches the MOWN Parition either via Regex or Exactly.\n",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "service",
            "in": "query",
            "description": "Matches the MOWN Service either via Regex or Exactly.\n",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "accountid",
            "in": "query",
            "description": "Matches the MOWN Account ID either via Regex or Exactly\n",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mownbase",
            "in": "query",
            "description": "Matches the MOWN Base URI fully (with regex or Exactly). Does not include\na matching of arguments\n",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "mownfull",
            "in": "query",
            "description": "Matches the MOWN full URI fully (with regex or Exactly). Includes a Matching of Arguments\n",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tagged",
            "in": "query",
            "description": "Matches an individual tag. Returns true if the tag is defined (even if the tag is set to something\nlike false). Specify using JQ style formatting. Uses the [JSON_EXISTS](https://mariadb.com/kb/en/json_exists/)\nfunctionality to do the JSON mapping. So **if you're looking for tag `loveit` you'd want to speicfy `$.loveit`**.\n",
            "required": false,
            "schema": {
              "type": "string"
            },
            "format": "jsonpath",
            "pattern": "^\\$\\."
          },
          {
            "name": "taggedtrue",
            "in": "query",
            "description": "Matches an individual tag. Returns true if the tag is defined as true Uses the \n[JSON_EXTRACT](https://mariadb.com/kb/en/json_extract/)\nfunctionality to do the JSON mapping. So **if you're looking for tag `loveit` you'd want to speicfy `$.loveit`**.\nWill match against [true](https://mariadb.com/kb/en/true-false/), a builtin boolean. So if you specify something\nlike `{\"loveit\" : \"false\"}` (Note the String not boolean) it will return true and if you have something like\n`{\"zerosareawesome\" : 0}` it should return false.\n",
            "required": false,
            "schema": {
              "type": "string"
            },
            "format": "jsonpath",
            "pattern": "^\\$\\.\""
          },
          {
            "name": "exact",
            "in": "query",
            "description": "Controls your other filters and flips them from regex matches (default) to exact matches\n",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                true,
                false
              ]
            }
          }
        ]
      }
    },
...
    "/hostcollections/{host_id}/": {
      "x-cached-length": "Every Midnight",
      "get": {
        "description": "Designed to grab the latest collections from the host. Grabs the fresh\nones as of yesterday Midnight. Only grabs one collection for each type/subtype\n",
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "tags": [
          "collections",
          "hosts"
        ],
        "parameters": [
          {
            "name": "host_id",
            "in": "path",
            "description": "The id of the host you wish to get data for.\n",
            "schema": {
              "type": "integer"
            },
            "required": true
          },
          {
            "name": "ctype",
            "in": "query",
            "description": "A regex to match for the collection_type. [PCRE](https://mariadb.com/kb/en/mariadb/regexp/) type\nregular expressions are accepted. Matched on the hostname column in the collection table.\n",
            "schema": {
              "type": "string"
            },
            "required": false
          },
          {
            "name": "exact",
            "in": "query",
            "description": "Controls your other filters and flips them from regex matches (default) to exact matches\n",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                true,
                false
              ]
            }
          },
          {
            "name": "value",
            "in": "query",
            "description": "Value of the collection You wish to search for. Regex by Default\n",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "csubtype",
            "in": "query",
            "description": "Value of the collection  subtype You wish to search for. Regex by Default\n",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  }
}